Beginning stages of Lost in the Woods

+++ trails/Models/Trail.cs
@@ -8,9 +8,9 @@ namespace trails.Models {
         public string Name { get; set; }
         [MinLength(10)]
         public string Description { get; set; }
-        public int Length { get; set; }
+        public decimal Length { get; set; }
         public int Elevation { get; set; }
-        public int Longitude { get; set; }
-        public int Latitude { get; set; }
+        public decimal Longitude { get; set; }
+        public decimal Latitude { get; set; }
     }
 }
